Change settings import
As far as I can tell the previous behaviour was to import 'settings'
from within the settings module. That meant it was importing
django.settings via socialfeedsparser.settings and ignoring all
configuration defined in socialfeedsparser.settings.

I'm not sure if that was ever working as expected.

I've changed that to import socialfeedsparser.settings itself which
resolves SOCIALFEEDSPARSER_TAG_TEMPLATE resolution errors I'm seeing,
and appears to fix the settings imports too.